Now designed for 1.21-1.21.3!
Sorry for taking so long with this. Life stuff is hard for me right now. Not much free time each day to work on this.
Wish I'd have more time each day to keep these up much quicker, but it's been real rough lately.
There's a lot of new stuff in this one though, so I hope you enjoy all that!