Now works for 1.21-1.21.4!
The Tricky Experiences update is out now, adding new EXP Jar variants for both early and late stages of the game, splash EXP Jars, autocrafting support, Sculk Catalyst interactions, new achievements and challenges, and a lot of Settings to mess aorund with! Details can be found in the Update Log.